# Conflicts:
#	docker/base/Dockerfile
#	public/html_header.php

+++ app/api/v3/Admin.php
@@ -1,22 +1,29 @@
 <?php
 /**
  * @author     Martin Høgh <mh@mapcentia.com>
- * @copyright  2013-2020 MapCentia ApS
+ * @copyright  2013-2021 MapCentia ApS
  * @license    http://www.gnu.org/licenses/#AGPL  GNU AFFERO GENERAL PUBLIC LICENSE 3
  *
  */
 
 namespace app\api\v3;
 
-ini_set('max_execution_time', 0);
+ini_set('max_execution_time', '0');
 
-use app\inc\Cache;
+use app\controllers\Mapcachefile;
+use app\controllers\Mapfile;
+use app\controllers\upload\Processqgis;
 use app\inc\Controller;
+use app\inc\Cache;
 use app\inc\Input;
+use app\inc\Model;
 use app\models\Database;
 use app\conf\App;
 use app\conf\Connection;
-use app\conf\migration\Sql;
+use app\migration\Sql;
+use app\models\Qgis;
+use PDOException;
+use Phpfastcache\Exceptions\PhpfastcacheInvalidArgumentException;
 
 
 /**
@@ -35,7 +42,7 @@ class Admin extends Controller
     }
 
     /**
-     * @return array
+     * @return array<mixed>
      *
      * @OA\Get(
      *   path="/api/v3/admin/mapfiles",
@@ -55,24 +62,27 @@ class Admin extends Controller
      *     )
      *   )
      * )
+     * @throws PhpfastcacheInvalidArgumentException
      */
     public function get_mapfiles(): array
     {
         $response = [];
         $database = new Database();
         $schemas = $database->listAllSchemas();
-        $mapfile = new \app\controllers\Mapfile();
-        if (!empty($schemas["data"])) foreach ($schemas["data"] as $schema) {
-            Connection::$param['postgisschema'] = $schema["schema"];
-            $res = $mapfile->get_index();
-            $response["data"][] = [$res[0]["ch"], $res[1]["ch"]];
+        $mapfile = new Mapfile();
+        if (!empty($schemas["data"])) {
+            foreach ($schemas["data"] as $schema) {
+                Connection::$param['postgisschema'] = $schema["schema"];
+                $res = $mapfile->get_index();
+                $response["data"][] = [$res[0]["ch"], $res[1]["ch"]];
+            }
         }
         $response["success"] = true;
         return $response;
     }
 
     /**
-     * @return array
+     * @return array<mixed>
      *
      * @OA\Get(
      *   path="/api/v3/admin/mapcachefile",
@@ -96,7 +106,7 @@ class Admin extends Controller
     public function get_mapcachefile(): array
     {
         $response = [];
-        $mapcachefile = new \app\controllers\Mapcachefile();
+        $mapcachefile = new Mapcachefile();
         $res = $mapcachefile->get_index();
         $response["data"] = $res["ch"];
         $response["success"] = true;
@@ -105,7 +115,7 @@ class Admin extends Controller
     }
 
     /**
-     * @return array
+     * @return array<mixed>
      *
      * @OA\Get(
      *   path="/api/v3/admin/qgisfiles",
@@ -128,12 +138,12 @@ class Admin extends Controller
      */
     public function get_qgisfiles(): array
     {
-        $qgis = new \app\models\Qgis();
+        $qgis = new Qgis();
         return $qgis->writeAll(Database::getDb());
     }
 
     /**
-     * @return array
+     * @return array<mixed>
      *
      * @OA\Get(
      *   path="/api/v3/admin/migrations",
@@ -161,8 +171,8 @@ class Admin extends Controller
 
         $arr = [Database::getDb(), "mapcentia", "gc2scheduler"];
         foreach ($arr as $db) {
-            \app\models\Database::setDb($db);
-            $conn = new \app\inc\Model();
+            Database::setDb($db);
+            $conn = new Model();
 
             switch ($db) {
                 case "mapcentia":
@@ -179,7 +189,7 @@ class Admin extends Controller
             foreach ($sqls as $sql) {
                 try {
                     $conn->execQuery($sql, "PDO", "transaction");
-                } catch (\PDOException $e) {
+                } catch (PDOException $e) {
                     $response["success"] = false;
                     $response["message"] = $e->getMessage();
                     return $response;
@@ -199,14 +209,8 @@ class Admin extends Controller
         return $response;
     }
 
-    public function get_reprocessqgis(): array
-    {
-        // TODO
-        // SELECT FROM settings.qgis_files WHERE db=....
-    }
-
     /**
-     * @return array
+     * @return array<mixed>
      *
      * DEPRECATED
      */
@@ -221,47 +225,52 @@ class Admin extends Controller
         } else {
             $files = glob(App::$param['path'] . "app/wms/qgsfiles/*.{qgs}", GLOB_BRACE);
         }
-        if (sizeof($files) == 0) {
+        if ($files && sizeof($files) == 0) {
             $response["code"] = 400;
             $response["success"] = false;
             $response["message"] = "No files";
             return $response;
         }
-        $qgis = new \app\models\Qgis();
-        $processqgis = new \app\controllers\upload\Processqgis();
+        $qgis = new Qgis();
+        $processqgis = new Processqgis();
 
-        usort($files, function ($a, $b) {
-            return filemtime($b) < filemtime($a);
-        });
+        if ($files) {
+            usort($files, function ($a, $b) {
+                return filemtime($b) < filemtime($a);
+            });
+            foreach ($files as $file) {
+                $bits1 = explode("/", $file);
+                $bits2 = explode("_", $bits1[$index]);
+                if ($bits2[0] == "parsed") {
+                    if (strlen($bits2[sizeof($bits2) - 1]) == 36) {
+                        $arr = [];
+                        $size = sizeof($bits2);
+                        for ($i = 1; $i < $size - 1; $i++) {
+                            $arr[] = $bits2[$i];
+                        }
+                        $orgFileName = implode("_", $arr) . ".qgs";
+                        $qgis->flagAsOld($bits1[$index]);
 
-        foreach ($files as $file) {
-            $bits1 = explode("/", $file);
-            $bits2 = explode("_", $bits1[$index]);
-            if ($bits2[0] == "parsed") {
-                if (strlen($bits2[sizeof($bits2) - 1]) == 36) {
-                    $arr = [];
-                    for ($i = 1; $i < sizeof($bits2) - 1; $i++) {
-                        $arr[] = $bits2[$i];
+                    } else {
+                        continue;
                     }
-                    $orgFileName = implode("_", $arr) . ".qgs";
-                    $qgis->flagAsOld($bits1[$index]);
-
                 } else {
-                    continue;
+                    $orgFileName = $bits1[$index];
                 }
-            } else {
-                $orgFileName = $bits1[$index];
-            }
 
-            $tmpFile = App::$param['path'] . "/app/tmp/" . Connection::$param["postgisdb"] . "/__qgis/" . $orgFileName;
+                $tmpFile = App::$param['path'] . "/app/tmp/" . Connection::$param["postgisdb"] . "/__qgis/" . $orgFileName;
 
-            if (copy($file, $tmpFile)) {
-                touch($tmpFile, filemtime($file));
-                $res = $processqgis->get_index($orgFileName);
-                if ($res["success"]) {
-                    unlink($file);
+                if (copy($file, $tmpFile)) {
+                    $time = filemtime($file);
+                    if ($time) {
+                        touch($tmpFile, $time);
+                    }
+                    $res = $processqgis->get_index($orgFileName);
+                    if ($res["success"]) {
+                        unlink($file);
+                    }
+                    $response["data"][] = $res["ch"];
                 }
-                $response["data"][] = $res["ch"];
             }
         }
         $response["success"] = true;
@@ -269,7 +278,7 @@ class Admin extends Controller
     }
 
     /**
-     * @return array
+     * @return array<mixed>
      *
      * @OA\Get(
      *   path="/api/v3/admin/schema",
@@ -293,11 +302,11 @@ class Admin extends Controller
     public function get_schema(): array
     {
         $admin = new \app\models\Admin();
-        return $admin->install();
+        return ["data" => $admin->install()];
     }
 
     /**
-     * @return array
+     * @return array<mixed>
      *
      * @OA\Get(
      *   path="/api/v3/admin/diskcleanup",
@@ -321,32 +330,40 @@ class Admin extends Controller
     public function get_diskcleanup(): array
     {
         $result = [];
-        function rrmdir($dir, &$result)
-        {
-            if (is_dir($dir)) {
-                $objects = scandir($dir);
+        $dirs = [App::$param["path"] . 'app/tmp'];
+        foreach ($dirs as $dir) {
+            $this->rrmdir($dir, $result);
+        }
+        return ["success" => true, "message" => "Unlinked files", "data" => $result];
+    }
+
+    /**
+     *
+     * @param string $dir
+     * @param array<string> $result
+     */
+    private function rrmdir(string $dir, array &$result): void
+    {
+        if (is_dir($dir)) {
+            $objects = scandir($dir);
+            if ($objects) {
                 foreach ($objects as $object) {
                     if ($object != "." && $object != ".." && $object != ".gitignore") {
                         if (is_dir($dir . "/" . $object) && !is_link($dir . "/" . $object)) {
-                            rrmdir($dir . "/" . $object, $result);
+                            $this->rrmdir($dir . "/" . $object, $result);
                         } else {
                             unlink($dir . "/" . $object);
                         }
                         $result[] = $dir . "/" . $object;
                     }
                 }
-                rmdir($dir);
             }
+            rmdir($dir);
         }
-        $dirs = [App::$param["path"] . 'app/tmp'];
-        foreach ($dirs as $dir) {
-            rrmdir($dir, $result);
-        }
-        return ["success" => true, "message" => "Unlinked files", "data" => $result];
     }
 
     /**
-     * @return array
+     * @return array<mixed>
      *
      * @OA\Get(
      *   path="/api/v3/admin/cachestats",
@@ -371,7 +388,7 @@ class Admin extends Controller
     }
 
     /**
-     * @return array
+     * @return array<mixed>
      *
      * @OA\Get(
      *   path="/api/v3/admin/cachecleanup",
